+++ test/e2e/scenario/rum/actions.scenario.ts
@@ -1,6 +1,5 @@
-import { createTest, html } from '../../lib/framework'
+import { createTest, html, waitForServersIdle } from '../../lib/framework'
 import { flushEvents } from '../../lib/sdkHelpers'
-import { waitForServersIdle } from '../../lib/servers'
 
 describe('user action collection', () => {
   createTest('track a click user action')
